requirements.txt
transformers
pillow
torch
torchvision
scikit-learn
rouge_score
Levenshtein
nltk
rouge-score
einops
timm
accelerate